+++ MOP/src/Systems/ConsoleCommands.cs
@@ -186,6 +186,9 @@ namespace MOP
                             say = "no u";
                             Process.Start("https://www.youtube.com/watch?v=dQw4w9WgXcQ");
                             break;
+                        case "a fact":
+                            say = "MOP 2.11 consists of 8340 lines of code!"
+;                            break;
                     }
 
                     ModConsole.Print($"< {say} >\n" +